pa-jail: improve EOF transmission
Sending EOF through a psuedo-TTY proves to be tricky. The current
solution causes this command to reliably hang

    $ seq 1 5000 > test
    $ pa-jail run --fg -f jfiles /jail jailuser wc < test

and this command to occasionally truncate output lines longer than 4096
characters:

    $ echo start > test; echo $(seq 1 5000) >> test; echo end >> test
    $ pa-jail run --fg -f jfiles /jail jailuser cat < test

As best as I can tell, the latter command's bug is a result of switching
the psuedo-TTY into canonical mode to transmit EOF from master to slave
*while* the slave is still transmitting data back to the master. If this
switch back to canonical mode occurs while a line has 4096 untransmitted
bytes, it will be truncated. (Switching the psuedo-TTY back into raw
mode immediately after transmitting EOF, for reasons I don't understand,
causes the EOF to be transmitted in raw mode, even when using
TCSADRAIN.)

The other way to send EOF down the line is to configure VMIN and VTIME
to cause read to return 0 (i.e., EOF) after 0.1s, which is what this
commit does. This keeps cat and wc happy. You'd think this would cause
interactive programs, like python and vim, to break, but it works on
everything I've tried.