pa-jail: only allocate a psuedo-tty if requested
Transferring regular files through a psuedo-TTY is at best a gross hack
and at worst incorrect. Before this commit, the following command would
occasionally truncate output lines longer than 4096 characters:

    $ seq 1 5000 > test
    $ pa-jail run --fg -f jfiles /jail jailuser cat < test

With this commit, pa-jail only installs a pseudo-TTY when explicitly
requested with the `-t` option. (This is a return to an earlier
behavior, which was removed in 21147b0.)

Psuedo-TTYs should only be requested when stdin (or the file specified
with `-i`) is a terminal or a FIFO. Requesting a TTY with a regular file
as input, like

    $ pa-jail run -t --fg -f jfiles /jail jailuser cat < test

will truncate long lines as before. SSH exhibits this same behavior if
you force it to allocate a psuedo-TTY:

    $ ssh -tt localhost cat < test