Optimize closestPoint for complex paths.
Create a “distance map”: an array of distance bounds for all
points with X value within 32 units. For instance, distmap[0]
is a lower bound on the distance from the search-point to
all points on the curve with X in the range
[distmap.xmin,distmap.xmin+distmap.width).

This allows us to make “big steps”. Consider path-length P,
which corresponds to point (X,Y). If no point in the bucket
corresponding to X is close, and no point in the adjacent buckets
is close either, then it is safe to advance P by distmap.width.

Also fix a boundary condition or two.