Use EventSource to collect output from runners.
This is less stressful on the machine -- less PHP.

Requires configuration. Set in conf/options.php the
$Opt["run_eventsourcedir"] variable to a directory;
"var/eventsource" is a good choice (relative directories are
interpreted relative to the site root). `pa-jail` will then
place a UNIX-domain socket in that directory on which it listens
for EventSource connections. Then configure NGINX to proxy
requests to `ROOT/runevents/v1/` to that directory. For example:

    location /grade/cs61-test/runevents/v1/ {
        alias /PATH/TO/ROOT/var/eventsource/;
        proxy_pass http://unix:$request_filename;
        proxy_http_version 1.1;
    }