Introduce random repository IDs (repogids).
A Peteramati installation depends on both database and file-system
data. Database IDs, especially repository IDs, are part of file-system
and Git repository information, such as git remotes and tags. This
is dangerous; if you change the database, then database and FS info
get out of sync. FS and git data should use unpredictable IDs,
different from installation to installation, so that changing the
database won't reliably confuse the file system.

This commit introduces repository GIDs, which are 7-character strings
starting with the cacheid, and supports upgrading git remotes and tags.